# v1/model/model_api_response.go
/*
 * Weather Aggregator API
 *
 * Aggregator API for collecting weather info of a location.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

// ApiResponse - Return the status response for a request or event
type ApiResponse struct {
	Code int32 `json:"code,omitempty"`

	Message string `json:"message,omitempty"`

	Details string `json:"details,omitempty"`
}