# v1/model/model_location.go
/*
 * Weather Aggregator API
 *
 * Aggregator API for collecting weather info of a location.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

// Location - Record of the specified landmark or location
type Location struct {
	Id string `json:"id,omitempty"`

	Longitude string `json:"longitude,omitempty" binding:"required"`

	Latitude string `json:"latitude,omitempty" binding:"required"`

	Name string `json:"name,omitempty" binding:"required"`

	Description string `json:"description,omitempty"`
}