# v1/model/model_weather.go
/*
 * Weather Aggregator API
 *
 * Aggregator API for collecting weather info of a location.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

import (
	"time"
)

// Weather - Collected weather information
type Weather struct {
	Id int64 `json:"id,omitempty"`

	AirTemperature float32 `json:"airTemperature,omitempty"`

	Humidity float32 `json:"humidity,omitempty"`

	Precipitation float32 `json:"precipitation,omitempty"`

	EventDateTime time.Time `json:"eventDateTime,omitempty"`

	Location Location `json:"location,omitempty" binding:"required"`
}